+++ docs/pages/webitel-ui/components/wt-action-bar/examples/example-slotted-table-action-bar.vue
@@ -1,6 +1,6 @@
 <template>
   <wt-action-bar
-    :actions="tableActionsOrder"
+    :include="tableActionsOrder"
     mode="table"
   >
     <template>