+++ docs/pages/webitel-ui/components/wt-indicator/examples/example-indicator-different-colors.vue
@@ -8,6 +8,9 @@ const colors = [
   'disabled',
   'break-out',
   'info',
+  'final-status',
+  'other-status',
+  'initial-status',
 ];
 </script>