+++ src/components/wt-datepicker/wt-datepicker.vue
@@ -14,7 +14,7 @@
         name="label"
         v-bind="{ label }"
       >
-        {{ label }}
+        {{ requiredLabel }}
       </slot>
     </wt-label>
     <vue-datepicker
@@ -80,9 +80,14 @@
 <script setup>
 import VueDatepicker from '@vuepic/vue-datepicker';
 import '@vuepic/vue-datepicker/dist/main.css';
-import { ref } from 'vue';
+import { computed, ref } from 'vue';
 
 const props = defineProps({
+
+  /**
+   * [`'date'`, `'datetime'`]
+   * */
+
   mode: {
     type: String,
     default: 'date',
@@ -114,9 +119,22 @@ const props = defineProps({
     type: String,
     default: 'en',
   },
+
+  /**
+   * Object with props, passed down to wt-label as props
+   */
+
   labelProps: {
     type: Object,
-    description: 'Object with props, passed down to wt-label as props',
+  },
+
+  /**
+   * Native input required attribute
+   */
+
+  required: {
+    type: Boolean,
+    default: false,
   },
 });
 const emit = defineEmits(['input']);
@@ -125,6 +143,10 @@ const isOpened = ref(false);
 const datepicker = ref(null); // template ref
 
 const isDateTime = props.mode === 'datetime';
+
+const requiredLabel = computed(() => {
+  return props.required ? `${props.label}*` : props.label;
+});
 </script>
 
 <style lang="scss">