+++ src/store/new/modules/tableStoreModule/tableStoreModule.js
@@ -1,3 +1,4 @@
+import deepCopy from 'deep-copy';
 import FilterEvent from '../../../../modules/Filters/enums/FilterEvent.enum.js';
 import {
   queryToSortAdapter,
@@ -11,6 +12,7 @@ const state = () => ({
   error: {},
   isLoading: false,
   isNextPage: false,
+  _resettable: {},
 });
 
 const getters = {
@@ -199,6 +201,7 @@ const actions = {
       throw err;
     } finally {
       await context.dispatch('LOAD_DATA_LIST');
+      await context.dispatch('SET_SELECTED', []);
 
       /* if no items on current page after DELETE, move to prev page [WTEL-3793] */
       if (!context.state.dataList.length && context.getters.FILTERS.page > 1) {
@@ -234,10 +237,21 @@ const actions = {
     context,
     payload,
   ) => context.dispatch('api/DELETE_ITEM', payload),
+
+  RESET_TABLE_STATE: async (context) => {
+    context.commit('RESET_TABLE_STATE');
+  },
+};
+
+const mutations = {
+  RESET_TABLE_STATE: (state) => {
+    Object.assign(state, deepCopy(state._resettable));
+  },
 };
 
 export default () => ({
   state: state(),
   getters,
   actions,
+  mutations,
 });